requirements.txt
google-generativeai
pytest
pandas
beautifulsoup4
requests
